Make address byteable
