+++ src/address.rs
@@ -1,3 +1,5 @@
+use crate::error::*;
+use cardano_multiplatform_lib::address::Address as CMLAddress;
 use serde::{Deserialize, Serialize};
 
 // TODO: Continue to hone this into a good API. I tried to make the Address generic, but it
@@ -24,6 +26,13 @@ impl Address {
             Address::Raw(inner) => inner,
             Address::Script(inner) => inner,
         }
+    }
+
+    pub fn bytes(&self) -> Result<Vec<u8>> {
+        // TODO: I'd rather not take a dep on CML here :(
+        let cml_address = CMLAddress::from_bech32(self.to_str())
+            .map_err(|e| Error::Address(format!("Error getting address bytes: {:?}", e)))?;
+        Ok(cml_address.to_bytes())
     }
 }
 
@@ -53,5 +62,23 @@ impl PolicyId {
                 }
             }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use cardano_multiplatform_lib::address::Address as CMLAddress;
+
+    #[test]
+    fn round_trip_address_bytes() {
+        let addr_str = "addr_test1qrksjmprvgcedgdt6rhg40590vr6exdzdc2hm5wc6pyl9ymkyskmqs55usm57gflrumk9kd63f3ty6r0l2tdfwfm28qs0rurdr";
+        let address = Address::new(addr_str);
+        let bytes = address.bytes().unwrap();
+        let new_cml_address = CMLAddress::from_bytes(bytes).unwrap();
+        let new_addr_str = new_cml_address
+            .to_bech32(Some("addr_test".to_string()))
+            .unwrap();
+        assert_eq!(addr_str, new_addr_str);
     }
 }
